requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
pymongo[srv]==3.12.0
python-dotenv==0.19.0
dnspython<2.0.0,>=1.16.0